Добавит слайс reviewsSlice
