+++ src/store/offers/offers-selectors.ts
@@ -4,6 +4,7 @@ import { State } from '../../types/state';
 import { selectCurrentCity } from '../city/city-selectors';
 import { filterOffersByCity, sortOffers } from '../../utils/utils';
 import { selectCurrentSortKey } from '../sort-key/sort-key-selectors';
+import { NEARBY_OFFERS_LIMITED } from '../../const';
 
 const selectOffers = (state: State): ShortOfferListType => state.offers;
 
@@ -19,9 +20,22 @@ const selectSortedOffers = createSelector(
 
 const selectIsOffersLoading = (state: State): boolean => state.isOffersLoading;
 
+const selectNearbyOffers = (state: State): ShortOfferListType =>
+  state.nearbyOffers;
+
+const selectIsNearbyOffersLoading = (state: State): boolean =>
+  state.isNearbyOffersLoading;
+
+const selectAdaptToNearbyOffers = createSelector(
+  [selectNearbyOffers],
+  (offers) => offers.slice(0, NEARBY_OFFERS_LIMITED)
+);
+
 export {
   selectOffers,
   selectIsOffersLoading,
   selectFilteredOffers,
   selectSortedOffers,
+  selectIsNearbyOffersLoading,
+  selectAdaptToNearbyOffers,
 };
