+++ src/store/reducer.ts
@@ -4,15 +4,7 @@ import {
   DEFAULT_CURRENT_CITY,
   DEFAULT_SORTING_KEY,
 } from '../const';
-import {
-  clearReviewsList,
-  setError,
-} from './actions';
-
-import {
-  submitReview,
-  uploadReviewsList,
-} from './api-actions';
+import { setError } from './actions';
 import { InitialState } from '../types/state';
 
 const initialState: InitialState = {
@@ -33,35 +25,9 @@ const initialState: InitialState = {
 };
 
 const reducer = createReducer(initialState, (builder) => {
-  builder
-    .addCase(uploadReviewsList.pending, (state) => {
-      state.isReviewsListLoading = true;
-    })
-    .addCase(uploadReviewsList.fulfilled, (state, action) => {
-      state.reviewsList = action.payload;
-      state.isReviewsListLoading = false;
-    })
-    .addCase(uploadReviewsList.rejected, (state) => {
-      state.reviewsList = [];
-      state.isReviewsListLoading = false;
-    })
-    .addCase(clearReviewsList, (state) => {
-      state.reviewsList = [];
-      state.isReviewsListLoading = false;
-    })
-    .addCase(submitReview.pending, (state) => {
-      state.isSubmitReviewLoading = true;
-    })
-    .addCase(submitReview.fulfilled, (state, action) => {
-      state.reviewsList = [...state.reviewsList, action.payload];
-      state.isSubmitReviewLoading = false;
-    })
-    .addCase(submitReview.rejected, (state) => {
-      state.isSubmitReviewLoading = false;
-    })
-    .addCase(setError, (state, action) => {
-      state.error = action.payload;
-    });
+  builder.addCase(setError, (state, action) => {
+    state.error = action.payload;
+  });
 });
 
 export { reducer };
